requirements.txt
pandas
numpy
matplotlib
pyspark
minio
boto3
dotenv
kafka-python
python-dotenv